# documentation/swagger.yaml
openapi: 3.0.0
info:
  title: MamaCheck API - Complete Reference
  description: |
    # MamaCheck - SMS-First Antenatal Care System

    Comprehensive REST API for pregnancy management, ANC reminders, and danger sign triage in Nigeria.

    ## Key Features
    - **Pregnancy Management**: Registration with OTP verification, timeline tracking
    - **ANC Reminders**: Daily scheduling at 07:00 WAT, personalized messages
    - **Danger Sign Triage**: Weekly check-ins with RED/YELLOW/GREEN outcomes
    - **CHEW Dashboard**: Real-time KPIs, women registry, red flag management
    - **Multi-language**: English, Pidgin, Yoruba, Hausa, Igbo
    - **SMS Integration**: Termii SMS gateway with DND compliance
    - **Visit Tracking**: Attendance marking with 10-minute undo window
    - **Reference Data**: LGA and PHC management with geolocation

    ## Security
    - JWT authentication for all protected endpoints
    - Role-based access (PATIENT, CHEW, SUPERVISOR, ADMIN)
    - OTP verification for phone number validation
    - Rate limiting on SMS and registration endpoints
    - BCRYPT password hashing in production

    ## SMS Triage Guide
    Women reply with symptom numbers:
    - **0** = No symptoms (GREEN)
    - **1** = Heavy bleeding (RED)
    - **2** = Severe headache (RED)
    - **3** = Swollen face/hands (RED)
    - **4** = Blurry vision (YELLOW)
    - **5** = Fever (YELLOW)
    - **6** = Reduced baby movement (YELLOW)
...
  version: "1.0.0"
...
servers:
...
paths:
...
  /pregnancies/register:
    post:
      tags:
        - Pregnancies
      summary: Register New Pregnancy
      description: |
        CHEW registers a pregnant woman. Collects:
        - Woman's name, phone, preferred language
        - Last Menstrual Period (LMP) or Expected Delivery Date (EDD)
        - Clinic name (PHC)
        - Trusted contact (optional)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregnancyRegistration"
      responses:
        "201":
          description: Pregnancy registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pregnancyId:
                    type: string
                  message:
                    type: string
        "400":
          description: Invalid OTP or missing fields
        "403":
          description: User is not a CHEW
  /pregnancies/chew/{chewId}:
    get:
      tags:
        - Pregnancies
      summary: Get CHEW's Pregnancies
      description: List all pregnancies registered under a CHEW
      security:
        - BearerAuth: []
      parameters:
        - name: chewId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of pregnancies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pregnancy"
  /pregnancies/{pregnancyId}:
    get:
      tags:
        - Pregnancies
      summary: Get Pregnancy Details
      security:
        - BearerAuth: []
      parameters:
        - name: pregnancyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pregnancy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pregnancy"
        "404":
          description: Pregnancy not found
  /pregnancies/{pregnancyId}/attended:
    post:
      tags:
        - Pregnancies
      summary: Mark ANC Visit Attended
      description: Record that a woman attended her scheduled ANC visit. Cancels reminders for that milestone.
      security:
        - BearerAuth: []
      parameters:
        - name: pregnancyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - milestoneNumber
              properties:
                milestoneNumber:
                  type: integer
                  minimum: 1
                  maximum: 8
                  example: 6
                  description: FMOH milestone number
                visitDate:
                  type: string
                  format: date
                  description: Date of visit (defaults to today)
      responses:
        "200":
          description: Visit marked as attended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: Pregnancy not found
  /pregnancies/{pregnancyId}/attended/undo:
    post:
      tags:
        - Pregnancies
      summary: Undo Visit Attendance
      description: |
        Reverse a visit attendance marking. 
        **Can only undo within 10-minute window after marking.**
        After 10 minutes, undo is not permitted.
      security:
        - BearerAuth: []
      parameters:
        - name: pregnancyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - milestoneNumber
              properties:
                milestoneNumber:
                  type: integer
                  example: 6
                reason:
                  type: string
                  example: "Marked in error"
      responses:
        "200":
          description: Attendance undone successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          description: Undo window expired (10 minutes) or no record found
  /pregnancies/{pregnancyId}/attendance-history:
    get:
      tags:
        - Pregnancies
      summary: Get Attendance History
      description: View all attendance records with undo availability info
      security:
        - BearerAuth: []
      parameters:
        - name: pregnancyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    visitWeek:
                      type: integer
                    action:
                      type: string
                      enum: [marked_attended, undone, unmarked]
                    markedAt:
                      type: string
                      format: date-time
                    canUndo:
                      type: boolean
                    undoWindowExpires:
                      type: string
                      format: date-time
  /pregnancies/{pregnancyId}/danger-reports:
    get:
      tags:
        - Pregnancies
      summary: Get Danger Reports
      description: Retrieve all danger/triage reports for a pregnancy
      security:
        - BearerAuth: []
      parameters:
        - name: pregnancyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of danger reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DangerReport"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [patient, chew, supervisor, admin]
        preferredLanguage:
          type: string
          enum: [en, pidgin, yo, ha, ig]
        trustedContact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
        consent:
          type: object
          properties:
            sms:
              type: boolean
        optOut:
          type: object
          properties:
            isOptedOut:
              type: boolean
    PregnancyRegistration:
      type: object
      required:
        - womanDetails
        - lmp
        - chewId
        - otp
      properties:
        womanDetails:
          type: object
          required:
            - name
            - phone
          properties:
            name:
              type: string
            phone:
              type: string
            preferredLanguage:
              type: string
              enum: [en, pidgin, yo, ha, ig]
              default: en
        lmp:
          type: string
          format: date
          example: "2025-09-18"
        edd:
          type: string
          format: date
        clinicName:
          type: string
          example: "Ungwan Rimi PHC"
        chewId:
          type: string
        otp:
          type: string
          minLength: 6
          maxLength: 6
        trustedContact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            relationship:
              type: string
            preferredLanguage:
              type: string
              enum: [en, pidgin, yo, ha, ig]
    Pregnancy:
      type: object
      properties:
        _id:
          type: string
        womanId:
          $ref: "#/components/schemas/User"
        gestationalWeek:
          type: integer
          minimum: 0
          maximum: 42
        status:
          type: string
          enum: [active, delivered, miscarried, transferred]
        lmp:
          type: string
          format: date
        edd:
          type: string
          format: date
        ancVisits:
          type: array
          items:
            type: object
            properties:
              weekNumber:
                type: integer
              status:
                type: string
                enum: [scheduled, attended, missed]
              scheduledDate:
                type: string
                format: date
              attendedDate:
                type: string
                format: date
    DangerReport:
      type: object
      properties:
        _id:
          type: string
        reportedSymptoms:
          type: array
          items:
            type: integer
            enum: [0, 1, 2, 3, 4, 5, 6, 7, 8]
        triageOutcome:
          type: string
          enum: [GREEN, YELLOW, RED]
        timestamp:
          type: string
          format: date-time
        requiresFollowup:
          type: boolean
        followup:
          type: object
          properties:
            status:
              type: string
              enum: [pending, in_progress, completed, escalated]
            outcome:
              type: string
              enum: [phone_contact, clinic_visit, referral_hospital]
...